requirements.txt
python-telegram-bot==20.7
nest_asyncio
gspread
oauth2client